requirements/base.txt
Django==1.5.4
Markdown==2.3.1
Pillow==2.1.0
South==0.8.2
argparse==1.2.1
distribute==0.6.24
django-crispy-forms==1.4.0
django-filebrowser==3.5.2
django-grappelli==2.4.6
djangorestframework==2.3.8
wsgiref==0.1.2
